p_12/requirements.txt
Django==4.0.6
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
pytz==2022.1
requests==2.28.1